feat: 전화번호 입력을 2단계에서 3단계(프로필)로 이동
