+++ src/features/auth/components/credential-step.tsx
@@ -1,25 +1,18 @@
 'use client';
 
 import { useEffect, useState } from 'react';
-import { Controller } from 'react-hook-form';
 
 import Link from 'next/link';
 
 import { useRegisterFormContext } from '@/features/auth/components/register-form-context-provider';
-import { PHONE_REGEX } from '@/features/auth/schemas/register';
 import { Button } from '@/shared/components/ui/button';
 import { Checkbox } from '@/shared/components/ui/checkbox';
 import { Form } from '@/shared/components/ui/form';
 import { Input } from '@/shared/components/ui/input';
 import { link } from '@/shared/constants/link';
 import { useCountdown } from '@/shared/hooks/use-countdown';
-import axios from 'axios';
 
-import {
-  useCheckEmailDuplicate,
-  useCheckPhoneNumberDuplicate,
-  useVerifyCode,
-} from '../services/query';
+import { useCheckEmailDuplicate, useVerifyCode } from '../services/query';
 
 const RESEND_COUNTDOWN = 30;
 const VERIFICATION_CODE_LENGTH = 6;
@@ -28,17 +21,8 @@ type CredentialStepProps = {
   onNext: () => void;
 };
 
-// 전화번호 자동 하이픈 포맷팅
-const formatPhoneNumber = (value: string) => {
-  const digits = value.replace(/\D/g, '').slice(0, 11);
-  if (digits.length <= 3) return digits;
-  if (digits.length <= 7) return `${digits.slice(0, 3)}-${digits.slice(3)}`;
-  return `${digits.slice(0, 3)}-${digits.slice(3, 7)}-${digits.slice(7)}`;
-};
-
 export const CredentialStep = ({ onNext }: CredentialStepProps) => {
   const [emailCodeVerified, setEmailCodeVerified] = useState(false);
-  const [isPhoneNumberChecked, setIsPhoneNumberChecked] = useState(false);
 
   const { countdown: resendCountdown, startCountdown } =
     useCountdown(RESEND_COUNTDOWN);
@@ -47,11 +31,6 @@ export const CredentialStep = ({ onNext }: CredentialStepProps) => {
     useCheckEmailDuplicate();
 
   const { mutate: verifyCode, isPending: isVerifyingCode } = useVerifyCode();
-
-  const {
-    mutate: checkPhoneNumberDuplicate,
-    isPending: isCheckingPhoneNumberDuplicate,
-  } = useCheckPhoneNumberDuplicate();
 
   const canResend = resendCountdown === null;
 
@@ -121,8 +100,7 @@ export const CredentialStep = ({ onNext }: CredentialStepProps) => {
     isAllRequiredTermsChecked &&
     emailCodeVerified &&
     isPasswordValid &&
-    isConfirmPasswordValid &&
-    isPhoneNumberChecked;
+    isConfirmPasswordValid;
 
   const verificationCodeInputValue = form.watch('verificationCode');
 
@@ -139,40 +117,6 @@ export const CredentialStep = ({ onNext }: CredentialStepProps) => {
       void form.trigger(['password', 'confirmPassword']);
     }
   }, [confirmPassword, form]);
-
-  // 전화번호 유효성 검사
-  const phoneNumberValue = form.watch('phoneNumber');
-  const isPhoneNumberValid = PHONE_REGEX.test(phoneNumberValue);
-
-  // 전화번호 중복 검사
-  const onCheckPhoneNumberButtonClick = () => {
-    if (isCheckingPhoneNumberDuplicate) return;
-
-    checkPhoneNumberDuplicate(
-      { phoneNumber: form.getValues('phoneNumber').replace(/-/g, '') },
-      {
-        onSuccess: () => {
-          setIsPhoneNumberChecked(true);
-          form.clearErrors('phoneNumber');
-        },
-        onError: (error) => {
-          if (
-            axios.isAxiosError(error) &&
-            error.response?.data?.code === 'PHONE_NUMBER_ALREADY_EXIST'
-          ) {
-            form.setError('phoneNumber', {
-              message: '이미 사용 중인 전화번호입니다.',
-            });
-          } else {
-            form.setError('phoneNumber', {
-              message: '전화번호 확인 중 오류가 발생했습니다.',
-            });
-          }
-          setIsPhoneNumberChecked(false);
-        },
-      }
-    );
-  };
 
   return (
     <div className="flex flex-col gap-8">
@@ -250,43 +194,6 @@ export const CredentialStep = ({ onNext }: CredentialStepProps) => {
         </Form.Control>
         <Form.ErrorMessage>
           {form.formState.errors.confirmPassword?.message}
-        </Form.ErrorMessage>
-      </Form.Item>
-      <Form.Item error={!!form.formState.errors.phoneNumber}>
-        <Form.Label>전화번호</Form.Label>
-        <div className="flex">
-          <Controller
-            name="phoneNumber"
-            control={form.control}
-            render={({ field }) => (
-              <Input
-                className="rounded-r-none border-r-0"
-                placeholder="010-0000-0000"
-                value={field.value}
-                onChange={(e) => {
-                  field.onChange(formatPhoneNumber(e.target.value));
-                  setIsPhoneNumberChecked(false);
-                  form.clearErrors('phoneNumber');
-                }}
-              />
-            )}
-          />
-          <Button
-            variant="secondary"
-            className="h-[56px] rounded-l-none"
-            type="button"
-            onClick={onCheckPhoneNumberButtonClick}
-            disabled={
-              isPhoneNumberChecked ||
-              !isPhoneNumberValid ||
-              isCheckingPhoneNumberDuplicate
-            }
-          >
-            중복 확인
-          </Button>
-        </div>
-        <Form.ErrorMessage>
-          {form.formState.errors.phoneNumber?.message}
         </Form.ErrorMessage>
       </Form.Item>
       <Checkbox.Group className="border-line-line1 flex flex-col gap-6 border-y py-6">
